+++ giggletech-router/src/main.rs
@@ -3,7 +3,7 @@
 // by Sideways
 // Based off OSC Async https://github.com/Frando/async-osc
 
-use async_osc::{prelude::*, OscPacket, OscType, Result};
+use anyhow::Result;
 use async_std::{stream::StreamExt, sync::Arc,};
 use std::sync::atomic::AtomicBool;
 
@@ -14,6 +14,7 @@ mod giggletech_osc;
 mod terminator;
 mod osc_timeout;
 mod handle_proximity_parameter;
+mod handle_osc_message;
 
 // TODO: clean up and refactor to avoid using as many .clone() statements
 
@@ -26,7 +27,7 @@ async fn main() -> Result<()> {
     let running = Arc::new(AtomicBool::new(false));
 
     // Rx/Tx Socket Setup
-    let mut rx_socket = giggletech_osc::setup_rx_socket(config.port_rx).await?;
+    let mut rx_socket = giggletech_osc::setup_rx_socket(&config.port_rx).await?;
 
     // Timeout
     for ip in config.headpat_device_uris.clone() {
@@ -37,41 +38,12 @@ async fn main() -> Result<()> {
         });
     }
     // Listen for OSC Packets
-    while let Some(packet) = rx_socket.next().await {
-        let (packet, _peer_addr) = packet?;
-
-        // Filter OSC Signals
-        match packet {
-            OscPacket::Bundle(_) => {}
-            OscPacket::Message(message) => {
-                let (address, osc_value) = message.as_tuple();
-                let value = match osc_value.first().unwrap_or(&OscType::Nil).clone().float() {
-                    Some(v) => v,
-                    None => continue,
-                };
-
-                // Max Speed Setting
-                if address == config.max_speed_parameter_address {
-                    data_processing::print_speed_limit(value);
-                    config.max_speed_float = value.max(config.max_speed_low_limit.clone());
-                } else {
-                    let index = config.proximity_parameters_multi.iter().position(|a| *a == address);
-                    if let Some(i) = index {
-                        handle_proximity_parameter::handle_proximity_parameter(
-                            running.clone(),
-                            &Arc::new(config.headpat_device_uris[i].clone()), 
-                            value, 
-                            config.max_speed_float, 
-                            config.min_speed_float, 
-                            config.speed_scale_float, 
-                            &config.proximity_parameters_multi[i]
-                        ).await?;
-                    } else {
-                        log::error!("TODO:")
-                    }
-                }
-            }
+    loop {
+        tokio::select! {
+            res = rx_socket.next() => {
+                handle_osc_message::handle_osc_message(res, &mut config, running.clone()).await?;
+            },
+            //TODO add twitch integration input
         }
     }
-    Ok(())
 }
